pyproject.toml
[project]
name = "arm4r"
version = "0.1.0"

dependencies=[
    "Pillow",
    "opencv-python",
    "tqdm", 
    "tensorboard",
    "diffusers",
    "scipy",
    "timm",
    "matplotlib", 
    "tyro", 
    "loralib",
    "h5py",
    "wandb",
    "mediapy",
    "kinpy",
    "imgviz",
    "transformers",
    "zarr",
    "open_clip_torch"
]

requires-python = ">=3.8"
[tool.setuptools.packages.find]
include = ["arm4r"]